decode: Don't write past the end of the input buffer
By changing the test program to allocate exactly the correct
amount of space for the decode buffer, we can see (using valgrind)
that the previous algorithm would (depending on the length) write
one byte past the end of the buffer.

Preserve the extra bits in a local variable to be added into the
next byte, rather than storing them in the next byte itself.